+++ python/VectorPartitioned.hpp
@@ -34,9 +34,9 @@ void init_VectorPartitioned(py::module& m)
         .def("iip", &GooseFEM::VectorPartitioned::iip, "Return prescribed degrees-of-freedom")
 
         .def(
-            "AsDofs",
+            "DofsFromParitioned",
             py::overload_cast<const xt::xtensor<double, 1>&, const xt::xtensor<double, 1>&>(
-                &GooseFEM::VectorPartitioned::AsDofs, py::const_),
+                &GooseFEM::VectorPartitioned::DofsFromParitioned, py::const_),
             "Set 'dofval",
             py::arg("dofval_u"),
             py::arg("dofval_p"))
@@ -70,17 +70,17 @@ void init_VectorPartitioned(py::module& m)
             py::arg("elemvec"))
 
         .def(
-            "AsNode",
+            "NodeFromPartitioned",
             py::overload_cast<const xt::xtensor<double, 1>&, const xt::xtensor<double, 1>&>(
-                &GooseFEM::VectorPartitioned::AsNode, py::const_),
+                &GooseFEM::VectorPartitioned::NodeFromPartitioned, py::const_),
             "Set 'nodevec",
             py::arg("dofval_u"),
             py::arg("dofval_p"))
 
         .def(
-            "AsElement",
+            "ElementFromPartitioned",
             py::overload_cast<const xt::xtensor<double, 1>&, const xt::xtensor<double, 1>&>(
-                &GooseFEM::VectorPartitioned::AsElement, py::const_),
+                &GooseFEM::VectorPartitioned::ElementFromPartitioned, py::const_),
             "Set 'elemvec",
             py::arg("dofval_u"),
             py::arg("dofval_p"))
@@ -116,133 +116,6 @@ void init_VectorPartitioned(py::module& m)
         .def("Copy_u", &GooseFEM::VectorPartitioned::Copy_u, "Copy iiu")
 
         .def("Copy_p", &GooseFEM::VectorPartitioned::Copy_p, "Copy iip")
-
-        // Overloaded method from Vector
-
-        .def("nelem", &GooseFEM::VectorPartitioned::nelem, "Number of element")
-
-        .def("nne", &GooseFEM::VectorPartitioned::nne, "Number of nodes per element")
-
-        .def("nnode", &GooseFEM::VectorPartitioned::nnode, "Number of nodes")
-
-        .def("ndim", &GooseFEM::VectorPartitioned::ndim, "Number of dimensions")
-
-        .def("ndof", &GooseFEM::VectorPartitioned::ndof, "Number of degrees-of-freedom")
-
-        .def("dofs", &GooseFEM::VectorPartitioned::dofs, "Return degrees-of-freedom")
-
-        .def(
-            "Copy",
-            &GooseFEM::Vector::Copy,
-            py::arg("nodevec_src"),
-            py::arg("nodevec_dest"))
-
-        .def(
-            "AsDofs",
-            py::overload_cast<const xt::xtensor<double, 2>&>(&GooseFEM::Vector::AsDofs, py::const_),
-            "Set 'dofval",
-            py::arg("nodevec"))
-
-        .def(
-            "AsDofs",
-            py::overload_cast<const xt::xtensor<double, 3>&>(&GooseFEM::Vector::AsDofs, py::const_),
-            "Set 'dofval",
-            py::arg("elemvec"))
-
-        .def(
-            "AsNode",
-            py::overload_cast<const xt::xtensor<double, 1>&>(&GooseFEM::Vector::AsNode, py::const_),
-            "Set 'nodevec",
-            py::arg("dofval"))
-
-        .def(
-            "AsNode",
-            py::overload_cast<const xt::xtensor<double, 3>&>(&GooseFEM::Vector::AsNode, py::const_),
-            "Set 'nodevec",
-            py::arg("elemvec"))
-
-        .def(
-            "AsElement",
-            py::overload_cast<const xt::xtensor<double, 1>&>(
-                &GooseFEM::Vector::AsElement, py::const_),
-            "Set 'elemvec",
-            py::arg("dofval"))
-
-        .def(
-            "AsElement",
-            py::overload_cast<const xt::xtensor<double, 2>&>(
-                &GooseFEM::Vector::AsElement, py::const_),
-            "Set 'elemvec",
-            py::arg("nodevec"))
-
-        .def(
-            "AssembleDofs",
-            py::overload_cast<const xt::xtensor<double, 2>&>(
-                &GooseFEM::Vector::AssembleDofs, py::const_),
-            "Assemble 'dofval'",
-            py::arg("nodevec"))
-
-        .def(
-            "AssembleDofs",
-            py::overload_cast<const xt::xtensor<double, 3>&>(
-                &GooseFEM::Vector::AssembleDofs, py::const_),
-            "Assemble 'dofval'",
-            py::arg("elemvec"))
-
-        .def(
-            "AssembleNode",
-            py::overload_cast<const xt::xtensor<double, 3>&>(
-                &GooseFEM::Vector::AssembleNode, py::const_),
-            "Assemble 'nodevec'",
-            py::arg("elemvec"))
-
-        .def(
-            "ShapeDofval",
-            &GooseFEM::Vector::ShapeDofval)
-
-        .def(
-            "ShapeNodevec",
-            &GooseFEM::Vector::ShapeNodevec)
-
-        .def(
-            "ShapeElemvec",
-            &GooseFEM::Vector::ShapeElemvec)
-
-        .def(
-            "ShapeElemmat",
-            &GooseFEM::Vector::ShapeElemmat)
-
-        .def(
-            "AllocateDofval",
-            py::overload_cast<>(&GooseFEM::Vector::AllocateDofval, py::const_))
-
-        .def(
-            "AllocateDofval",
-            py::overload_cast<double>(&GooseFEM::Vector::AllocateDofval, py::const_))
-
-        .def(
-            "AllocateNodevec",
-            py::overload_cast<>(&GooseFEM::Vector::AllocateNodevec, py::const_))
-
-        .def(
-            "AllocateNodevec",
-            py::overload_cast<double>(&GooseFEM::Vector::AllocateNodevec, py::const_))
-
-        .def(
-            "AllocateElemvec",
-            py::overload_cast<>(&GooseFEM::Vector::AllocateElemvec, py::const_))
-
-        .def(
-            "AllocateElemvec",
-            py::overload_cast<double>(&GooseFEM::Vector::AllocateElemvec, py::const_))
-
-        .def(
-            "AllocateElemmat",
-            py::overload_cast<>(&GooseFEM::Vector::AllocateElemmat, py::const_))
-
-        .def(
-            "AllocateElemmat",
-            py::overload_cast<double>(&GooseFEM::Vector::AllocateElemmat, py::const_))
 
         .def("__repr__", [](const GooseFEM::VectorPartitioned&) {
             return "<GooseFEM.VectorPartitioned>";